[claude] L2-P5: code remediation batch — JSF-182/190/202 + JPL-25 fixes, -Werror gate (bench_sim)
Clears all known mechanical/decidable JSF/P10 violations dispositioned in the
2026-06-23 decision-lock (6f2cf6b). All fix-oriented; no new accepted deviations
beyond the recorded CAST-1/CAST-2.

- mechanical: shadow x1, #pragma once->#ifndef x3, offsetof->ptr-span x2, (void*)0->nullptr x24
- JSF-202 float==: CAL_STATUS_WMM_SET bit + has_default_location (no flash-layout change)
- JSF-182 casts: evt_cast<E> centralizes QEvt downcasts (CAST-2), CRC const void*,
  buffer overlays -> static_cast (Exc1), MPU cast -> uintptr_t (CAST-1)
- JSF-190 continue: all 28 guard-clause continues inverted (12 in eskf flight math)
- JPL-25 >6 params: MavCmdParams, calibration->cal_vec3_t, CombinatorSpec, FlightEntryLayout
- gate: -Wshadow/-Wfloat-equal flipped to hard errors (-Werror), all sites cleared
- generator: generate_profile.py emits #ifndef + has_default_location; new
  CODING_STANDARDS "Auto-Generated Code" rule (mission_profile_data.h drift)

Verified: host ctest 857/857; WSL firmware clean -Werror rebuild, 0 warnings;
bit-exact equivalence vs HEAD on both flight-critical math paths (eskf+ud_factor
continue inversions, calibration Vec3 refactor); bench_sim 2/2 PASS on flashed
firmware (vehicle flight v0.16.0 kmenu, COM7, sensors GO).

+++ src/main.cpp
@@ -207,7 +207,7 @@ static void init_fault_recovery() {
     // Shared fault handlers (OPT-IVP-01) + MPU stack guard.
     exception_set_exclusive_handler(HARDFAULT_EXCEPTION, memmanage_fault_handler);
     exception_set_exclusive_handler(MEMMANAGE_EXCEPTION, memmanage_fault_handler);
-    mpu_setup_stack_guard(reinterpret_cast<uint32_t>(&__StackBottom));
+    mpu_setup_stack_guard(reinterpret_cast<uintptr_t>(&__StackBottom));
 
     // R-25-exec: latches probe-armed test mode + clears the SRAM magic.
     rc::test_mode_init();
